GLOBAL-862 Удаление чата без выхода с экрана

+++ features/profile-guest/profile-guest-impl/src/main/java/io/fasthome/fenestram_messenger/profile_guest_impl/data/repo_impl/ProfileGuestImpl.kt
@@ -5,7 +5,7 @@ import io.fasthome.fenestram_messenger.profile_guest_impl.domain.repo.ProfileGue
 import io.fasthome.fenestram_messenger.util.callForResult
 
 class ProfileGuestImpl(private val profileGuestService: ProfileGuestService) : ProfileGuestRepo {
-    override suspend fun deleteChat(id: Int) = callForResult {
+    override suspend fun deleteChat(id: Long) = callForResult {
         profileGuestService.deleteChat(id)
     }
 }
