+++ features/profile-guest/profile-guest-impl/src/main/java/io/fasthome/fenestram_messenger/profile_guest_impl/domain/logic/DeleteChatUseCase.kt
@@ -5,5 +5,5 @@ import io.fasthome.fenestram_messenger.util.onSuccess
 
 class DeleteChatUseCase(private val profileGuestRepo: ProfileGuestRepo) {
 
-    suspend fun deleteChat(id: Int) = profileGuestRepo.deleteChat(id).onSuccess { }
+    suspend fun deleteChat(id: Long) = profileGuestRepo.deleteChat(id).onSuccess { }
 }
